Added comments

+++ Cviceni_3/src/Cviceni_3/Predmet.java
@@ -1,10 +1,11 @@
 package Cviceni_3;
 
 public interface Predmet {
+
     int pointsToCredit = 15;
 
     String getName();
-    int getPoints();
+    double getPoints();
     void getEval();
     String getMark();
